Move EntitySection into Navigation namespace
