+++ visitz/VisitzModel/Models/Navigation/EntityNavItem.cs
@@ -1,5 +1,3 @@
-using VisitzModel.Models.EntityTypes;
-
 namespace VisitzModel.Models.Navigation;
 
 public class EntityNavItem : NavItem
